move spectrum_plot.py in tool. modified plotSpectrum() and add function plotNormalizedSpectrum()
